Set null-count in linked_column_view conversion operator
