+++ cpp/include/cudf/detail/utilities/linked_column.hpp
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022, NVIDIA CORPORATION.
+ * Copyright (c) 2022-2023, NVIDIA CORPORATION.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,9 +16,11 @@
 
 #pragma once
 
+#include <cudf/column/column_view.hpp>
 #include <cudf/table/table_view.hpp>
 
-#include <thrust/iterator/transform_iterator.h>
+#include <memory>
+#include <vector>
 
 namespace cudf::detail {
 
@@ -35,30 +37,11 @@ struct linked_column_view : public column_view_base {
   linked_column_view(linked_column_view const&) = delete;
   linked_column_view& operator=(linked_column_view const&) = delete;
 
-  linked_column_view(column_view const& col) : linked_column_view(nullptr, col) {}
+  linked_column_view(column_view const& col);
 
-  linked_column_view(linked_column_view* parent, column_view const& col)
-    : column_view_base(col), parent(parent)
-  {
-    children.reserve(col.num_children());
-    std::transform(
-      col.child_begin(), col.child_end(), std::back_inserter(children), [&](column_view const& c) {
-        return std::make_shared<linked_column_view>(this, c);
-      });
-  }
+  linked_column_view(linked_column_view* parent, column_view const& col);
 
-  operator column_view() const
-  {
-    auto child_it = thrust::make_transform_iterator(
-      children.begin(), [](auto const& c) { return static_cast<column_view>(*c); });
-    return column_view(this->type(),
-                       this->size(),
-                       this->head(),
-                       this->null_mask(),
-                       UNKNOWN_NULL_COUNT,
-                       this->offset(),
-                       std::vector<column_view>(child_it, child_it + children.size()));
-  }
+  operator column_view() const;
 
   linked_column_view* parent;  //!< Pointer to parent of this column. Nullptr if root
   LinkedColVector children;
@@ -70,11 +53,6 @@ struct linked_column_view : public column_view_base {
  * @param table table of columns to convert
  * @return Vector of converted linked_column_views
  */
-inline LinkedColVector table_to_linked_columns(table_view const& table)
-{
-  auto linked_it = thrust::make_transform_iterator(
-    table.begin(), [](auto const& c) { return std::make_shared<linked_column_view>(c); });
-  return LinkedColVector(linked_it, linked_it + table.num_columns());
-}
+LinkedColVector table_to_linked_columns(table_view const& table);
 
 }  // namespace cudf::detail
